requirements.txt
openai
httpx
pypdf
python-pptx
python-dotenv
